Switch to cleaner pillar structure
This commit allows users to specify sysctl parameters in an easier
format than the one that was used before. It moves settings out of the
'lookup' pillar namespace in a backward compatible manner and is
modelled after the old 'params2' approach.

It also introduces a simplified scheme, in which parameters can be
specified as key-value pairs, which makes for cleaner pillars if there
is no need to specify a custom configuration file.